feat : 이주의 술 dto에 컬럼 추가
양조장, 평가 수 컬럼 추가

+++ src/main/java/com/example/soonsul/user/repository/PersonalEvaluationRepository.java
@@ -17,7 +17,7 @@ import java.util.Optional;
 public interface PersonalEvaluationRepository extends JpaRepository<PersonalEvaluation,Long> {
 
     Optional<PersonalEvaluation> findByUserAndLiquor(User user, Liquor liquor);
-    Long countByLiquor(Liquor liquor);
+    Integer countByLiquor(Liquor liquor);
     Integer countByUser(User user);
 
     @Query(nativeQuery = true,
